Homework Assignment #17: The Notifications API
Details:
 
As you've browsed around the web, you've undoubtedly encountered prompts like this:



This is an example of how the WHATWG Notifications API works. Users are initially prompted for permission to send notifications, and if they consent, the requester can send browser-notifications to them at any time.

For this assignment, you are to use MDN as a resource to learn everything you can about this useful API. Start here:

https://developer.mozilla.org/en-US/docs/Web/API/Notifications_API/Using_the_Notifications_API

Once you've familiarized yourself with the API, build a standalone webpage that acts as a simple notification-based alarm clock. It should allow a user to:

1. Select the time and date they want an "alarm" (notification)

2. Receive that notification at the requested time/date (the alarm should just be a simple notification, no sound needed)

3. Add multiple alarms

4. Delete or edit any alarms that haven't been triggered yet

5. View all past alarms that have been triggered


Extra Credit:

Use the localstorage API to allow the user-data and alarm-data to persist between page refreshes.